requirements.txt
asgiref==3.8.1
autopep8==2.3.2
coverage==7.10.6
Django==5.2.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.4.0
iniconfig==2.1.0
mypy==1.16.1
mypy_extensions==1.1.0
packaging==25.0
parameterized==0.9.0
pathspec==0.12.1
pillow==11.3.0
pluggy==1.6.0
pycodestyle==2.13.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
